refactor/drawerCategories
